PYIC-3265 Rename the uk.gov.di.ipv.core.processjourneystep.statemachine.responses package to uk.gov.di.ipv.core.processjourneystep.statemachine.stepresponses
