+++ lambdas/process-journey-event/src/main/java/uk/gov/di/ipv/core/processjourneyevent/ProcessJourneyEventHandler.java
@@ -26,8 +26,8 @@ import uk.gov.di.ipv.core.processjourneyevent.statemachine.StateMachineInitializ
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.StateMachineNotFoundException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownEventException;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.exceptions.UnknownStateException;
-import uk.gov.di.ipv.core.processjourneyevent.statemachine.responses.JourneyContext;
-import uk.gov.di.ipv.core.processjourneyevent.statemachine.responses.PageResponse;
+import uk.gov.di.ipv.core.processjourneyevent.statemachine.stepresponses.JourneyContext;
+import uk.gov.di.ipv.core.processjourneyevent.statemachine.stepresponses.PageResponse;
 import uk.gov.di.ipv.core.processjourneyevent.statemachine.states.BasicState;
 
 import java.io.IOException;
